# docs/apispec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "User Management API",
    "description": "API for user management",
    "contact": {
      "name": "Muhammad Fathi",
      "email": "contact.fathisiddiqi@gmail.com",
      "url": "https://fathisiddiqi.github.io"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "API Server for local development"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Authentication API"],
        "summary": "Login user",
        "description": "Login an user",
        "requestBody": {
          "description": "Request body for login",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "required": ["true"]
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 8,
                    "required": ["true"]
                  }
                }
              },
              "examples": {
                "abdul": {
                  "description": "Request body for login example",
                  "value": {
                    "email": "abdulqadir@gmail.com",
                    "password": "loremipsum78"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Response body for login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "success": {
                    "description": "Examples success login user",
                    "value": {
                      "code": 200,
                      "status": "OK",
                      "data": {
                        "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiQWJkdWwgUWFkaXIiLCJwYXNzd29yZCI6ImxvcmVtaXNwdW03OCIsImlhdCI6MTUxNjIzOTAyMn0.rXwu6WUbJjqKtLmiJJfJxgDgp1OXtkeSl9dRdJSYEyM",
                        "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiQWJkdWwgUWFkaXIiLCJwYXNzd29yZCI6ImxvcmVtaXNwdW03OCIsImlhdCI6MTUxNjIzOTAyMn0.rXwu6WUbJjqKtLmiJJfJxgDgp1OXtkeSl9dRdJSYEyM"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Response body not authorized user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                },
                "examples": {
                  "failed": {
                    "description": "Example for failed login user",
                    "value": {
                      "code": 401,
                      "status": "UNAUTHORIZED",
                      "error": {
                        "name": "ErrUNAUTHORIZED",
                        "message": "wrong username and password"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Authentication API"],
        "summary": "Register user",
        "description": "Register user",
        "requestBody": {
          "description": "Request body for register",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "minLength": 1,
                    "required": ["true"]
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 1,
                    "required": ["true"]
                  },
                  "birthOfDate": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "required": ["true"]
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 8,
                    "required": ["true"]
                  }
                }
              },
              "examples": {
                "abdul": {
                  "description": "Request body for login example",
                  "value": {
                    "firstName": "Abdul",
                    "lastName": "Qadir",
                    "email": "abdulqadir@gmail.com",
                    "birthOfDate": "2000-10-10",
                    "password": "loremipsum78"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Response body for register",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string"
                        },
                        "refreshToken": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "success": {
                    "description": "Examples success register user",
                    "value": {
                      "code": 200,
                      "status": "OK",
                      "data": {
                        "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiQWJkdWwgUWFkaXIiLCJwYXNzd29yZCI6ImxvcmVtaXNwdW03OCIsImlhdCI6MTUxNjIzOTAyMn0.rXwu6WUbJjqKtLmiJJfJxgDgp1OXtkeSl9dRdJSYEyM",
                        "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiQWJkdWwgUWFkaXIiLCJwYXNzd29yZCI6ImxvcmVtaXNwdW03OCIsImlhdCI6MTUxNjIzOTAyMn0.rXwu6WUbJjqKtLmiJJfJxgDgp1OXtkeSl9dRdJSYEyM"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Responses for failed register user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                },
                "examples": {
                  "failed": {
                    "description": "Response for failed register user",
                    "value": {
                      "code": 400,
                      "status": "BAD REQUEST",
                      "error": {
                        "name": "ErrValidation",
                        "message": "`firstName` value cannot be empty"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "tags": ["Authentication API"],
        "summary": "Forgot password",
        "description": "Forgot password",
        "requestBody": {
          "description": "Request body for forgot password",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              },
              "examples": {
                "abdul": {
                  "description": "Example for abdul in request body",
                  "value": {
                    "email": "abdulqadir@gmail.com"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Response success for forgot password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "format": "int32"
                    },
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "success": {
                    "description": "Example for response success forgot password",
                    "value": {
                      "code": 200,
                      "status": "OK",
                      "data": {
                        "message": "Email reset password sent successfully"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Failed response for forgot password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Err"
                },
                "examples": {
                  "failed": {
                    "description": "Failed response for forgot password",
                    "value": {
                      "code": 404,
                      "status": "NOT FOUND",
                      "errors": [
                        {
                          "name": "ErrNotFound",
                          "message": "user not found"
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "schemas": {
...
      "Err": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "error": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}